Prepare for ACL support.
This changes access checks from a straight uid compare into a call to
job_access(), which checks whether a particular user has access to the
job.  For now, it only gives access to a user operating on their own
jobs, but this provides a framework for ACLs to be implemented.